(use "git restore --staged <file>..." to unstage) 	modified:   Shared/Handler/SessionHandler.swift

+++ Shared/Handler/SessionHandler.swift
@@ -43,7 +43,7 @@ class SessionHandler: ObservableObject {
     @Published var isShowingAlert = false
     
     // MARK: - Settings
-    @AppStorage("fontSize") var fontSize: Int = 12
+    @AppStorage("fontSize") var fontSize: CGFloat = 12
 
     // MARK: - Initialization
